backend/requirements.txt
Django==4.2.2
Flask==2.2.2
requests==2.28.1     
gunicorn==20.1.0      
djangorestframework==3.14.0  
Werkzeug==2.2.2
